requirements.txt
# 合同对比助手 - 依赖包
# Contract Diff Tool - Requirements

# GUI 框架
PyQt6>=6.4.0

# Word 文档处理
python-docx>=0.8.11

# 差异对比算法
diff-match-patch>=20200713

# Excel 报告导出
openpyxl>=3.0.10

# PDF 处理（可选，用于 PDF 文档对比）
pdfplumber>=0.9.0
pdf2docx>=0.5.6

# 网络请求（用于自动更新）
requests>=2.28.0

# 打包工具
pyinstaller>=5.0

# 测试框架（开发用）
pytest>=7.0.0
